+++ src/charactersData/characters/Layla.ts
@@ -1,0 +1,46 @@
+import {Character, Vision} from '../../const/consts';
+import {
+  Books,
+  BossMaterials,
+  BossSkillMaterial,
+  Flower,
+  Gem,
+  Materials
+} from '../materials/materialNames';
+import {fillAscensionMaterials, fillTalentMaterials} from '../../utils/utils';
+
+const Layla = {
+  name: Character.LAYLA,
+  vision: Vision.CRYO,
+  ascension_materials: fillAscensionMaterials({
+    gems: {
+      1: Gem.CRYO.SHIVADA_JADE_SLIVER,
+      2: Gem.CRYO.SHIVADA_JADE_FRAGMENT,
+      3: Gem.CRYO.SHIVADA_JADE_CHUNK,
+      4: Gem.CRYO.SHIVADA_JADE_GEMSTONE
+    },
+    materials: {
+      1: Materials.DIVINING_SCROLL,
+      2: Materials.SEALED_SCROLL,
+      3: Materials.FORBIDDEN_CURSE_SCROLL,
+    },
+    bossMaterial: BossMaterials.PERPETUAL_CALIBER,
+    specialty: Flower.NILOTPALA_LOTUS
+  }),
+  rarity: 4,
+  talent_materials: fillTalentMaterials({
+    books: {
+      1: Books.TEACHINGS_OF_INGENUITY,
+      2: Books.GUIDE_TO_INGENUITY,
+      3: Books.PHILOSOPHIES_OF_INGENUITY,
+    },
+    materials: {
+      1: Materials.DIVINING_SCROLL,
+      2: Materials.SEALED_SCROLL,
+      3: Materials.FORBIDDEN_CURSE_SCROLL,
+    },
+    bossMaterial: BossSkillMaterial.MIRROR_OF_MUSHIN
+  })
+}
+
+export default Layla